More LatencyParameters
