+++ hystrix-sample/src/main/java/io/acari/pojo/SessionParameters.java
@@ -1,0 +1,14 @@
+package io.acari.pojo;
+
+public class SessionParameters {
+    private Long sessionId;
+
+
+    public Long getSessionId() {
+        return sessionId;
+    }
+
+    public void setSessionId(Long sessionId) {
+        this.sessionId = sessionId;
+    }
+}
